backend/requirements.txt
#backend
bcrypt==3.2.2
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
cryptography==37.0.4
PyMySQL==1.0.2
mysqlclient==2.1.1

#great api
boto3

#ai
keras-models
numpy
pillow